+++ src/Tools/Helpers.ts
@@ -27,3 +27,30 @@ export function pushDefined<T>(array: T[], ...items: (T | undefined | null)[]) {
 export function delay(ms: number) {
     return new Promise( (resolve) => setTimeout(resolve, ms) );
 }
+
+
+/**
+ * Returns a string representing the difference between two times with format `hh:mm:ss.fff`
+ * @param start Start time
+ * @param end End time
+ */
+export function timeDiffString(start: Date, end: Date): string {
+    let remain = end.getTime() - start.getTime();
+    // millis
+    const millis = remain % 1000;
+    remain = Math.floor(remain / 1000);
+    // seconds
+    const seconds = remain % 60;
+    remain = Math.floor(remain / 60);
+    // minutes
+    const minutes = remain % 60;
+    remain = Math.floor(remain / 60);
+    // hours
+    const hours = Math.floor(remain / 60);
+    // string
+    const fff = `${millis}`.padStart(3, '0');
+    const ss = `${seconds}`.padStart(2, '0');
+    const mm = `${minutes}`.padStart(2, '0');
+    const hh = `${hours}`.padStart(2, '0');
+    return `${hh}:${mm}:${ss}.${fff}`;
+}