+++ CHANGELOG.md
@@ -5,8 +5,8 @@ All notable changes to the "vscode-brautomationtools" extension will be document
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-Add new but unreleased features, fixes... here during development
+## [0.0.6] - 2022-01-17
+This release adds some small usability features.
 
 ### Added
 - The build of an Automation Studio project now shows timestamps and the total duration of the build